publish non-generic QuantizeProcessor
